+++ lib/app/modules/camera_pick/views/camera_pick_view.dart
@@ -63,7 +63,8 @@ class CameraPickView extends StatelessWidget {
                           icon: const Icon(Icons.clear,
                               color: Colors.white),
                           onPressed: () {
-                            Get.back();
+                            // Get.back();
+                            Navigator.pop(context);
                           },
                         ),
                         Expanded(